Added Triggers in DB and added exception handling

+++ backend/src/main/java/com/example/librarySystem/publisher/PublisherController.java
@@ -1,6 +1,7 @@
 package com.example.librarySystem.publisher;
 
 import lombok.AllArgsConstructor;
+import org.springframework.dao.DataAccessException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -18,10 +19,14 @@ public class PublisherController {
     @PutMapping("/edit/{name}")
     public ResponseEntity<String> editPublisher(@PathVariable String name, @RequestBody Publisher publisher) {
         try {
-            System.out.println(publisher);
+            if(!publisherService.publisherExists(name))
+                return new ResponseEntity<>("Publisher doesn't exist", HttpStatus.BAD_REQUEST);
             publisherService.editPublisher(name, publisher);
             return new ResponseEntity<>("Publisher edited successfully", HttpStatus.OK);
         } catch (Exception e) {
+            if (publisherService.PublisherExistsByPhone(publisher.getPhone()))
+                return new ResponseEntity<>("Publisher Phone Already Exists", HttpStatus.BAD_REQUEST);
+            System.out.println(e.getMessage());
             return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
@@ -29,11 +34,14 @@ public class PublisherController {
     @DeleteMapping("/delete/{name}")
     public ResponseEntity<String> deletePublisher(@PathVariable String name) {
         try{
+            if(!publisherService.publisherExists(name))
+                return new ResponseEntity<>("Publisher Doesn't Exist", HttpStatus.BAD_REQUEST);
             publisherService.deletePublisher(name);
-            return new ResponseEntity<>("Publisher was deleted successfully.", HttpStatus.CREATED);
+            return new ResponseEntity<>("Publisher was deleted successfully.", HttpStatus.OK);
         }
         catch(Exception e){
-            return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
+            System.out.println(e.getMessage());
+            return new ResponseEntity<>("Publisher can't be deleted.", HttpStatus.BAD_REQUEST);
         }
     }
 
@@ -57,9 +65,13 @@ public class PublisherController {
         try {
             publisherService.createPublisher(publisher);
             return new ResponseEntity<>("Publisher was created successfully.", HttpStatus.CREATED);
-        } catch (Exception e) {
-            System.out.println(e.getMessage());
-            return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
+        } catch (DataAccessException dae) {
+            if(publisherService.publisherExists(publisher.getName()))
+                return new ResponseEntity<>("Publisher Name Already Exists", HttpStatus.BAD_REQUEST);
+            else if (publisherService.PublisherExistsByPhone(publisher.getPhone()))
+                return new ResponseEntity<>("Publisher Phone Already Exists", HttpStatus.BAD_REQUEST);
+            System.out.println("Message = " + dae.getMessage());
+            return new ResponseEntity<>("Bad Request", HttpStatus.BAD_REQUEST);
         }
     }
